Исходя из спецификации 1. Сменил тип аннотации для параметров в контроллере; 2. Сменил типы запросов в контроллере; 3. Сменил тип данных некоторых полей в DTO;

+++ src/main/java/com/example/diploma/dto/Comment.java
@@ -10,7 +10,7 @@ public class Comment {
     private int author;
     private String authorImage;
     private String authorFirstName;
-    private int createdAt;
+    private long createdAt;
     private int pk;
     private String text;
 }
